+++ website/src/components/resume/Education.js
@@ -1,7 +1,7 @@
  
 import React from 'react';
 import './Education.css';
-import resume from './../assets/data/resume.json'
+import resume from './../../assets/data/resume.json'
 import settings from "./../../website-setting.json"