+++ release-notes.md
@@ -1,4 +1,4 @@
 
 ### Changes:
 
-- New repository
+- Move assets folder out of components